Initial commit to create skeleton for the startblock plugin
The startblock plugin will be mostly a frontend plugin, and it will
have one submit rule (block changes with hashtag "BLOCK_SUBMISSION")
which will be implemented in a follow-up.

This plugin should be used in conjuction with the startblock bot
(to be implemented).

This is mostly just copied from I00156c89.

Change-Id: I5988168a37e8a03e22c59f1b368f56dc53eca071